+++ src/index.ts
@@ -18,3 +18,4 @@ if (
 }
 
 export * from './BasisTheory';
+export * from './types';